requirements.txt
playwright
BeautifulSoup4
tinkoff-investments